Break compute functions up into submodules
