+++ desc/compute/__init__.py
@@ -1,5 +1,5 @@
 from .data_index import data_index
-from .compute_funs import (
+from ._core import (
     compute_flux_coords,
     compute_toroidal_flux,
     compute_toroidal_coords,
@@ -12,6 +12,9 @@ from .compute_funs import (
     compute_jacobian,
     compute_covariant_metric_coefficients,
     compute_contravariant_metric_coefficients,
+    compute_geometry,
+)
+from ._field import (
     compute_contravariant_magnetic_field,
     compute_covariant_magnetic_field,
     compute_magnetic_field_magnitude,
@@ -19,12 +22,10 @@ from .compute_funs import (
     compute_magnetic_tension,
     compute_B_dot_gradB,
     compute_contravariant_current_density,
-    compute_force_error,
-    compute_boozer_coords,
-    compute_quasisymmetry_error,
-    compute_energy,
-    compute_geometry,
 )
+from ._qs import compute_boozer_coords, compute_quasisymmetry_error
+from ._equil import compute_force_error, compute_energy
+
 
 # defines the order in which objective arguments get concatenated into the state vector
 arg_order = ("R_lmn", "Z_lmn", "L_lmn", "Rb_lmn", "Zb_lmn", "p_l", "i_l", "Psi")
